requirements.txt
dvc[ssh]==3.16.0
jupyterlab==4.0.5
matplotlib==3.7.2
opencv-python==4.8.0.76
pandas==2.0.3
lightning[extra]==2.0.8
albumentations==1.3.1
pytest==7.4.0
torchvision==0.15.2
pydantic==2.1.0
omegaconf==2.3.0
scikit-multilearn==0.2.0
timm==0.9.5
clearml==1.12.2
onnxruntime==1.16.0
onnx==1.14.1